Support libc backend
