+++ src/backend/linux_raw/fs/mod.rs
@@ -1,6 +1,4 @@
 pub(crate) mod dir;
 pub(crate) mod makedev;
-#[cfg(all(linux_raw, feature = "fs"))]
-pub(crate) mod raw_dir;
 pub(crate) mod syscalls;
 pub(crate) mod types;
